Move HeaderAdSlot up and out of Header folder
